FINAL VERSION 1.2.1

+++ src/Modelos/Modeloproductoexterno.java
@@ -37,7 +37,7 @@ public class Modeloproductoexterno extends Controladorproductoexterno{
         try (ResultSet rs = ps.executeQuery(sSQL)) {
             while (rs.next()) {          
                  columna[0] = rs.getString(1);                
-                columna[1] = (int) rs.getFloat(2);                              
+                columna[1] = rs.getFloat(2);                              
                 modeloT.addRow(columna);
             }
         }
@@ -59,7 +59,7 @@ public class Modeloproductoexterno extends Controladorproductoexterno{
         try (ResultSet rs = ps.executeQuery(sSQL)) {
             while (rs.next()) {
                  columna[0] = rs.getString(1);                
-                columna[1] = (int) rs.getFloat(2);
+                columna[1] =  rs.getFloat(2);
                 modeloT.addRow(columna);
             }
         }
@@ -81,7 +81,7 @@ public class Modeloproductoexterno extends Controladorproductoexterno{
         try (ResultSet rs = ps.executeQuery(sSQL)) {
             while (rs.next()) {
                  columna[0] = rs.getString(1);                
-                columna[1] = (int) rs.getFloat(2);
+                columna[1] =  rs.getFloat(2);
                 modeloT.addRow(columna);
             }
         }
